requirements.txt
cachelib==0.9.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
Flask-Caching==2.0.2
Flask-Cors==3.0.10
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pycryptodome==3.17
requests==2.28.2
six==1.16.0
urllib3==1.26.14
Werkzeug==2.2.3
